various minor refactoring
